% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.dynamodb-package}
\alias{aws.dynamodb-package}
\alias{aws.dynamodb}
\title{aws.dynamodb}
\description{
A simple client package for AWS Dynamo DB.
}
\details{
A simple client package for AWS Dynamo DB.
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/index.html}{API Guide}
}
\seealso{
\code{\link{create_table}}, \code{\link{get_table}}, \code{\link{get_item}}, \code{\link{put_item}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
