% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\alias{update_table}
\title{Create/Update Dynamo DB Table}
\usage{
create_table(table, attributes, primary_key, sort_key = NULL,
  read_capacity = 5, write_capacity = 5, ...)

update_table(table, read_capacity = 5, write_capacity = 5, ...)
}
\arguments{
\item{table}{A character string specifying the table name.}

\item{attributes}{A named list of key-value pairs, where the key is the attribute name and value is one of \dQuote{S} (string), \dQuote{N} (number), or \dQuote{B} (binary). Users only need to specify the attribute(s) used for the primary key.}

\item{primary_key}{The primary key for the database. This should be the name of one of the attributes specified in \code{attributes}.}

\item{read_capacity}{A number specifying read capacity units (i.e., number of reads per second). See References.}

\item{write_capacity}{A number specifying write capacity units (i.e., number of reads per second). See References.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\description{
Create or update a Dynamo DB Table
}
\examples{
\dontrun{
  tab <- create_table(
    table = "Music",
    attributes = list(Artist = "S", SongTitle = "S"),
    primary_key = "Artist",
    sort_key = "SongTitle"
  )

  # cleanup
  delete_table(tab)
}
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_CreateTable.html}{API Guide: CreateTable}
  \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/ProvisionedThroughput.html}{Provisioned Throughput}
  \href{https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.CoreComponents.html#HowItWorks.CoreComponents.PrimaryKey}{Primary Key}
}
