% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_item.R
\name{delete_item}
\alias{delete_item}
\title{Get Dynamo DB Item}
\usage{
delete_item(table, item, ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{item}{A list of key-value pairs. If the table only has a primary key, this should specify the primary key attribute name and value for the desired item. If a composite primary key is used, then both attribute names and values must be specified.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Get an item from a Dynamo DB tables
}
\examples{
\dontrun{
  tab <- create_table(
    table = "Music",
    attributes = list(Artist = "S"),
    primary_key = "Artist"
  )

  # put item
  put_item("Music", list(Artist = "No One You Know", SongTitle = "Call Me Today"))

  # get item
  get_item("Music", list(Artist = "No One You Know"))

  # delete item
  delete_item("Music", list(Artist = "No One You Know"))

  # cleanup
  delete_table(tab)
}
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DeleteItem.html}{API Guide: DeleteItem}
}
