% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_table.R
\name{delete_table}
\alias{delete_table}
\title{Delete a Dynamo DB Table}
\usage{
delete_table(table, ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Delete a Dynamo DB table
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DeleteTable.html}{API Guide: DeleteTable}
}
