% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get Dynamo DB Tables}
\usage{
get_table(table, ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Get a description of a Dynamo DB tables
}
