% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_backups.R
\name{list_table_backups}
\alias{list_table_backups}
\title{List Dynamo DB Backups}
\usage{
list_table_backups(marker = NULL, n = 100, ...)
}
\arguments{
\item{marker}{A character string containing the first backup table to return (contained as an attribute in a previous call to this function)}

\item{n}{A number between 1 and 100 specifying the total number of tables to return}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list of backup tables.
}
\description{
Retreive a paginated list of backup tables
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_ListBackups.html}{API Guide: ListBackups}
}
