% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List Dynamo DB Tables}
\usage{
list_tables(marker = NULL, n = 100, ...)
}
\arguments{
\item{marker}{A character string containing the first table to return (contained as an attribute in a previous call to this function)}

\item{n}{A number between 1 and 100 specifying the total number of tables to return}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list of table names.
}
\description{
Retreive a paginated list of Dynamo DB tables
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_ListTables.html}{API Guide: ListTables}
}
