% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_item.R
\name{put_item}
\alias{put_item}
\alias{update_item}
\title{Put an item}
\usage{
put_item(table, item, return_value = c("NONE", "ALL_OLD"), ...)

update_item(table, item, return_value = c("NONE", "ALL_OLD"), ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{item}{A list of key-value pairs.}

\item{return_value}{A character string specifying whether to return previous values of the item.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Put an item into a Dynamo DB database
}
\examples{
\dontrun{
  tab <- create_table(
    table = "Music",
    attributes = list(Artist = "S"),
    primary_key = "Artist"
  )

  # put item
  put_item("Music", list(Artist = "No One You Know", SongTitle = "Call Me Today"))

  # cleanup
  delete_table(tab)
}
}
