% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_item.R
\name{update_item}
\alias{update_item}
\title{Update Dynamo DB Item}
\usage{
update_item(table, key, item, return_value = c("NONE", "ALL_OLD"), ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{key}{A list of key-value pairs.  If the table only has a primary key, this should specify the primary key attribute name and value for the desired item. If a composite primary key is used, then both attribute names and values must be specified.}

\item{item}{A list of key-value pairs.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Update an item from a Dynamo DB tables
}
\examples{
\dontrun{
  tab <- create_table(
    table = "Music",
    attributes = list(Artist = "S"),
    primary_key = "Artist"
  )
  
  # put item
  put_item("Music", list(Artist = "No One You Know", SongTitle = "Call Me Yesterday"))

  # get item
  get_item("Music", list(Artist = "No One You Know"))
  
  # update item
  update_item("Music", key = list(Artist = "No One You Know", Songtitle = "Call Me Today"))
  
  # get item
  get_item("Music", list(Artist = "No One You Know"))

  # cleanup
  delete_table(tab)
}
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateItem.html}{API Guide: UpdateItem}
}
