% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.dynamodb-package}
\alias{aws.dynamodb-package}
\alias{aws.dynamodb}
\title{aws.dynamodb}
\description{
A simple client package for AWS Dynamo DB.
}
\details{
A simple client package for AWS Dynamo DB.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
