% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_table.R
\name{create_table_backup}
\alias{create_table_backup}
\alias{restore_table_backup}
\title{Backup a Dynamo DB Table}
\usage{
create_table_backup(table, backup, ...)

restore_table_backup(table, backup, ...)
}
\arguments{
\item{table}{A character string specifying the table name, or an object of class \dQuote{aws_dynamodb_table}.}

\item{backup}{For creating: a character string specifying the backup table name. For restoring: a backup table ARN.}

\item{\dots}{Additional arguments passed to \code{\link{dynamoHTTP}}.}
}
\value{
A list.
}
\description{
Create a backup of a Dynamo DB table, or restore a table from a backup
}
\references{
\href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_CreateBackup.html}{API Guide: CreateBackup}
  \href{https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_RestoreTableFromBackup.html}{API Guide: RestoreTableFromBackup}
}
\seealso{
\code{\link{list_backups}}
}
